backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
geoalchemy2==0.14.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation and Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

# Monitoring and Logging
structlog==23.2.0

# CORS (FastAPI built-in via starlette)
# python-cors is not needed; FastAPI uses starlette's CORSMiddleware